Important Notice: > Imported Mirror
> Imported ParrelSync (let's you clone your project so that you can test the networking stuff without building and running project every time- you can do it straight from the editor)

> Created PlayerController than encompasses NetworkBehaviour (old input system). Mirror is more friendly with the old input system.

> Additions to ChickenRed prefab (adding network identity components) 

> Creation of NetworkManager

Note: not all buttons such as "End Host" "End Client" work when testing... not yet. This is so because functions like onClientDisconnect, onServerDisconnect, etc... have not been implemented. You can setup a host game and then join as a client from the clone you generate from ParrelSync, but you won't be able to disconnect (for now just press the play button again to end the game).